16,17-02-23 mapping

+++ src/main/java/dao/RequestDao/CourseRequestDao.java
@@ -6,24 +6,24 @@ import lombok.Data;
 public class CourseRequestDao {
 	
 	private String name;
-	private int department_id;
-	private int student_id;
+	private Long department_id;
+	private Long student_id;
 	public String getName() {
 		return name;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getDepartment_id() {
+	public Long getDepartment_id() {
 		return department_id;
 	}
-	public void setDepartment_id(int department_id) {
+	public void setDepartment_id(Long department_id) {
 		this.department_id = department_id;
 	}
-	public int getStudent_id() {
+	public Long getStudent_id() {
 		return student_id;
 	}
-	public void setStudent_id(int student_id) {
+	public void setStudent_id(Long student_id) {
 		this.student_id = student_id;
 	}
 }
